fix: Gestion des dépendances circulaires avec @Lazy - TODO: sauvegarde des filtres en bdd
