+++ src/main/java/fr/initiativedeuxsevres/trouve_ton_match/controller/UtilisateurController.java
@@ -28,33 +28,11 @@ import java.util.stream.Collectors;
 @RequestMapping("/creationCompte")
 public class UtilisateurController {
 
-    private final ParrainService parrainService;
-
-    private final PorteurService porteurService;
-
     private final UtilisateurService utilisateurService;
 
     private final TypeAccompagnementService typeAccompagnementService;
 
     private final SecteurReseauService secteurReseauService;
-
-//    private final UtilisateurMapper utilisateurMapper;
-
-//    private final ParrainMapper parrainMapper;
-
-//    private final PorteurMapper porteurMapper;
-
-
-    // @Autowired
-    // public UtilisateurController(ParrainService parrainService, PorteurService
-    // porteurService, UtilisateurService utilisateurService,
-    // TypeAccompagnementService typeAccompagnementService) {
-    // this.parrainService = parrainService;
-    // this.porteurService = porteurService;
-    // this.utilisateurService = utilisateurService;
-    // this.typeAccompagnementService = typeAccompagnementService;
-    // this.secteurReseauService = secteurReseauService;
-    // }
 
     @PostMapping(value = "/createutilisateur", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Utilisateur> createUtilisateur(@RequestBody UtilisateurDto utilisateurDto) {
@@ -100,112 +78,6 @@ public class UtilisateurController {
         return utilisateur
                 .map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
-    }
-
-//    @PostMapping(value = "/completercompteparrain", produces = MediaType.APPLICATION_JSON_VALUE)
-//    public ResponseEntity<Parrain> completercompteparrain(@RequestBody ParrainDto parrainDto) {
-//
-//        System.out.println("Données reçues : " + parrainDto);
-//
-//        Parrain parrain = parrainService.findById(parrainDto.getIdUtilisateur());
-//
-//        // Vérifier si le parrain existe dans la base
-//        if (parrain == null) {
-//            // Si le parrain n'existe pas, retourner un objet Parrain vide
-//            Parrain emptyParrain = new Parrain();
-//            emptyParrain.setIdUtilisateur(parrainDto.getIdUtilisateur());
-//            System.out.println("Parrain introuvable, renvoi d'un objet vide.");
-//            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(emptyParrain);
-//        }
-//
-//        System.out.println("Compte complété");
-//        System.out.println("id :" + parrainDto.getIdUtilisateur());
-//
-//        // Mise à jour des données
-//        parrain.setPresentationParcours(parrainDto.getPresentationParcours());
-//        parrain.setBranchesReseau(parrainDto.getBranchesReseau());
-//        parrain.setDomainesExpertise(parrainDto.getDomainesExpertise());
-//        parrain.setSecteurGeographique(parrainDto.getSecteurGeographique());
-//        parrain.setDisponibilites(parrainDto.getDisponibilites());
-//
-//        Parrain updatedParrainEntity = parrainMapper.toEntity(parrainDto);  // Utiliser un Mapper pour convertir le DTO en entité
-//        Parrain updatedParrain = parrainService.save(updatedParrainEntity);  // Sauvegarder l'entité dans la base
-//
-//
-////        Parrain updatedParrain = parrainService.save(parrain);
-//        return ResponseEntity.ok(updatedParrain);
-//    }
-
-    @PostMapping (value = "/completercompteparrain", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Parrain> completercompteparrain(@RequestBody ParrainDto parrainDto) {
-
-//        System.out.println("Données reçues : " + parrainDto);
-//
-//        // Récupérer le parrain existant par son ID
-//        Parrain parrainExistant = parrainService.findById(parrainDto.getIdUtilisateur());
-//        System.out.println("Récupérer le parrain existant par son ID dans /completercompteparrain : " + parrainDto);
-//        System.out.println("Récupérer le NOM du parrain existant par son ID dans /completercompteparrain : " + parrainDto.getNomUtilisateur());
-//
-//        // Vérifier si le parrain existe dans la base
-//        if (parrainExistant == null) {
-//            // Si le parrain n'existe pas, retourner un objet Parrain vide
-//            Parrain emptyParrain = new Parrain();
-//            emptyParrain.setIdUtilisateur(parrainDto.getIdUtilisateur());
-//            System.out.println("Parrain introuvable, renvoi d'un objet vide.");
-//            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(emptyParrain);
-//        }
-//
-//        // Récupérer la liste des Accompagnements et Secteurs
-//        List<TypeAccompagnementDto> accompagnements = typeAccompagnementService.findAll();
-//        List<SecteurReseauDto> secteursReseaux = secteurReseauService.findAll();
-//
-//        // Mapper le DTO en entité, en incluant les accompagnements et secteurs
-//        Parrain updatedParrainEntity =  parrainMapper.toEntity(parrainDto, parrainExistant, accompagnements, secteursReseaux);
-//        System.out.println("Entité mise à jour après le mapping : " + updatedParrainEntity);
-//        System.out.println("parrainDto dans Mapper le DTO en entité: " + parrainDto);
-//        System.out.println("updatedParrainEntity via Mapper le DTO en entité: " + updatedParrainEntity);
-
-        // Sauvegarder l'entité
-        Parrain updatedParrain = parrainService.save(parrainDto);
-
-        // Retourner la réponse avec l'entité mise à jour
-        return ResponseEntity.ok(updatedParrain);
-    }
-
-    @PostMapping(value = "/completercompteporteur", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Porteur> completercompteporteur(@RequestBody PorteurDto porteurDto) {
-
-//        System.out.println("Données reçues : " + porteurDto);
-//
-//        Porteur porteurExistant = porteurService.findById(porteurDto.getIdUtilisateur());
-//
-//        // Vérifier si le porteur existe dans la base
-//        if (porteurExistant == null) {
-//            // Si le porteur n'existe pas, retourner un objet Porteur vide
-//            Porteur emptyPorteur = new Porteur();
-//            emptyPorteur.setIdUtilisateur(porteurDto.getIdUtilisateur());
-//            System.out.println("Porteur introuvable, renvoi d'un objet vide.");
-//            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(emptyPorteur);
-//        }
-//
-//        System.out.println("Compte complété");
-//        System.out.println("id :" + porteurDto.getIdUtilisateur());
-//
-//        // Récupérer la liste des Accompagnements et Secteurs
-//        List<TypeAccompagnementDto> accompagnements = typeAccompagnementService.findAll();
-//        List<SecteurReseauDto> secteursReseaux = secteurReseauService.findAll();
-//
-//        // Mapper le DTO en entité, en incluant les accompagnements et secteurs
-//        Porteur updatedPorteurEntity =  porteurMapper.toEntity(porteurDto, porteurExistant, accompagnements, secteursReseaux);
-//        System.out.println("Entité mise à jour après le mapping : " + updatedPorteurEntity);
-//        System.out.println("parrainDto dans Mapper le DTO en entité: " + porteurDto);
-//        System.out.println("updatedParrainEntity via Mapper le DTO en entité: " + updatedPorteurEntity);
-
-        // Sauvegarder l'entité
-        Porteur updatedPorteur = porteurService.save(porteurDto);
-
-        // Retourner la réponse avec l'entité mise à jour
-        return ResponseEntity.ok(updatedPorteur);
     }
 
     @PostMapping(value = "/accompagnementutilisateur", produces = MediaType.APPLICATION_JSON_VALUE)
